edit004 finish log design

+++ ramcube.h
@@ -4,7 +4,7 @@
 
 void ramcube_adjust_memcached_settings(struct settings *);
 void ramcube_init(struct event_base *, const struct settings *);
-int ramcube_process_commands(conn *, void *, const size_t, char *left_com);
+ulong ramcube_process_commands(conn *, void *, const size_t, char *left_com);
 int ramcube_server_proxy_new(conn *);
 ulong ramcube_post_set_data(conn *);
